+++ lib/bds/lib/model/operation-types.ts
@@ -54,6 +54,7 @@ export enum OperationTypes {
   AddKafka = "ADD_KAFKA",
   RemoveKafka = "REMOVE_KAFKA",
   ExecuteBootstrapScript = "EXECUTE_BOOTSTRAP_SCRIPT",
+  OdhServiceCertificateUpdate = "ODH_SERVICE_CERTIFICATE_UPDATE",
 
   /**
    * This value is used if a service returns a value for this enum that is not recognized by this