+++ lib/identitydomains/lib/model/authentication-factors-remover.ts
@@ -175,6 +175,19 @@ export interface AuthenticationFactorsRemover {
    *  - idcsSearchable: false
    */
   "type"?: AuthenticationFactorsRemover.Type;
+  /**
+   * Specifies the token to be removed
+   * <p>
+   **SCIM++ Properties:**
+   *  - type: string
+   *  - multiValued: false
+   *  - required: false
+   *  - mutability: readWrite
+   *  - returned: request
+   *  - uniqueness: none
+   *  - idcsSearchable: false
+   */
+  "token"?: string;
 }
 
 export namespace AuthenticationFactorsRemover {