+++ lib/identitydomains/lib/model/freeform-tags.ts
@@ -53,6 +53,7 @@ export interface FreeformTags {
    **SCIM++ Properties:**
    *  - caseExact: false
    *  - required: true
+   *  - idcsReturnEmptyWhenNull: true
    *  - mutability: readWrite
    *  - returned: default
    *  - type: string